remove lodash dependencies
