+++ index.js
@@ -1,147 +1,10 @@
 'use strict';
 
-const isObject = require('lodash.isobject');
-const isFunction = require('lodash.isfunction');
-const defaults = require('lodash.defaults');
+require('./monkey-patch');
+
 const mpath = require('mpath');
 const debug = require('debug')('mapper');
-const reg_dollar_exist = /\$/;
-const isNum = /^\d+$/;
 
-// wait https://github.com/aheckmann/mpath/pull/6
-function K(v) {
-  return v;
-}
-mpath.set = function(path, val, o, special, map, _copying, workWithArray) {
-  var lookup;
-
-  if ('function' == typeof special) {
-    if (special.length < 2) {
-      map = special;
-      special = undefined;
-    } else {
-      lookup = special;
-      special = undefined;
-    }
-  }
-
-  map || (map = K);
-
-  var parts = 'string' == typeof path
-    ? path.split('.')
-    : path;
-
-  if (!Array.isArray(parts)) {
-    throw new TypeError('Invalid `path`. Must be either string or array');
-  }
-
-  if (null == o) return;
-
-  // the existance of $ in a path tells us if the user desires
-  // the copying of an array instead of setting each value of
-  // the array to the one by one to matching positions of the
-  // current array.
-  var copy = _copying || reg_dollar_exist.test(path)
-    , obj = o
-    , part;
-
-  for (var i = 0, len = parts.length - 1; i < len; ++i) {
-    part = parts[i];
-
-    if ('$' == part) {
-      if (i == len - 1) {
-        break;
-      } else {
-        continue;
-      }
-    }
-
-    if (Array.isArray(obj) && !isNum.test(part)) {
-      var paths = parts.slice(i);
-
-      workWithArray = true;
-
-      if (!copy && Array.isArray(val)) {
-        for (var j = 0; j < obj.length && j < val.length; ++j) {
-          // assignment of single values of array
-          exports.set(paths, val[j], obj[j], special || lookup, map, copy);
-        }
-      } else {
-        for (var j = 0; j < obj.length; ++j) {
-          // assignment of entire value
-          exports.set(paths, val, obj[j], special || lookup, map, copy);
-        }
-      }
-      return;
-    }
-
-    if (!obj[part]) {
-      if (parts[i + 1] && parts[i + 1] == '$' && !parts[i + 2]) {
-        obj[part] = [];
-      }
-
-      if (!workWithArray && parts[i + 1] != '$') {
-        obj[part] = {};
-      }
-    }
-
-    if (lookup) {
-      obj = lookup(obj, part);
-    } else {
-      obj = special && obj[special]
-        ? obj[special][part]
-        : obj[part];
-    }
-
-    if (!obj) return;
-  }
-
-  // process the last property of the path
-  part = parts[len];
-
-  // use the special property if exists
-  if (special && obj[special]) {
-    obj = obj[special];
-  }
-
-  // set the value on the last branch
-  if (Array.isArray(obj) && !isNum.test(part)) {
-    if (!copy && Array.isArray(val)) {
-      for (var item, j = 0; j < obj.length && j < val.length; ++j) {
-        item = obj[j];
-        if (item) {
-          if (lookup) {
-            lookup(item, part, map(val[j]));
-          } else {
-            if (item[special]) item = item[special];
-            item[part] = map(val[j]);
-          }
-        }
-      }
-      // push element in array
-    } else if (Array.isArray(obj) && part == '$') {
-      obj.push(map(val));
-    } else {
-      for (var j = 0; j < obj.length; ++j) {
-        item = obj[j];
-        if (item) {
-          if (lookup) {
-            lookup(item, part, map(val));
-          } else {
-            if (item[special]) item = item[special];
-            item[part] = map(val);
-          }
-        }
-      }
-    }
-  } else {
-    if (lookup) {
-      lookup(obj, part, map(val));
-    } else {
-      obj[part] = map(val);
-    }
-  }
-};
 
 /**
  * mapper
@@ -188,7 +51,7 @@ mpath.set = function(path, val, o, special, map, _copying, workWithArray) {
  */
 class Mapper {
   constructor(map, options) {
-    if (!isObject(map)) {
+    if (!map || typeof map !== 'object') {
       throw new TypeError('Map is not object');
     }
 
@@ -203,11 +66,11 @@ class Mapper {
       throw new TypeError('Map is empty');
     }
 
-    this.options = defaults(options, {
+    this.options = Object.assign({}, {
       skipError: false,
       skipFields: '',
       delimiter: ' '
-    });
+    }, options);
   }
 
   /**
@@ -287,7 +150,7 @@ class Mapper {
    * @param {Object} map { where: what }
    */
   injector(map) {
-    if (map && isObject(map)) {
+    if (Boolean(map) && typeof map === 'object') {
       this.setValue(map, this._dst);
     }
   }
@@ -302,7 +165,7 @@ class Mapper {
     const context = {src: this._src, dst: this._dst, handler};
 
     if (value != null) {
-      const func = (isFunction(handler)) ? handler : this.bridge;
+      const func = (typeof handler === 'function') ? handler : this.bridge;
 
       return new Promise(resolve => resolve(func.call(context, value)))
         .then(result => this.injector(result))
@@ -348,18 +211,17 @@ class Mapper {
    * @returns {*}
    */
   transfer(src, dst) {
-    if (!isObject(src) || !isObject(dst)) {
+    if (!src || typeof src !== 'object' || !dst || typeof dst !== 'object') {
       return Promise.reject(new TypeError('Source or destination is not object'));
     }
 
     this._src = src;
     this._dst = dst;
 
-    const promise = Object
+    return Object
       .keys(this._map)
-      .reduce((accumulator, source_path) => this.reduceHandler(accumulator, this._map[source_path], source_path), Promise.resolve());
-
-    return promise.then(() => dst);
+      .reduce((accumulator, source_path) => this.reduceHandler(accumulator, this._map[source_path], source_path), Promise.resolve())
+      .then(() => dst);
   }
 }
